renamde/reorder components
