+++ frontend/package.json
@@ -14,6 +14,7 @@
     "react-dropdown": "^1.11.0",
     "react-icons": "^4.11.0",
     "react-responsive": "^9.0.2",
+    "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
