+++ frontend/src/Components/Base/Base.module.scss
@@ -1,4 +1,4 @@
-@use '../variables';
+@use '../../variables';
 
 .base {
     min-height: calc(100vh - 0vh);
