+++ frontend/src/Components/Header/Header.modules.scss
@@ -1,0 +1,18 @@
+@use '../../variables.scss';
+
+nav {
+    height: variables.$headerbar;
+
+    background-color: white;
+
+    position: sticky;
+    top: 0;
+
+    z-index: 1000;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    filter: drop-shadow(0 0 5px variables.$shadow-colour);
+}
