+++ frontend/src/Components/League/League.module.scss
@@ -1,0 +1,1 @@
+@use '../../variables';
