+++ frontend/src/Components/LeagueTable/LeagueTable.module.scss
@@ -130,6 +130,7 @@ $breakpoints: variables.$resolution-breakpoints;
             // display: inline;
             img {
                 height: 1.75rem;
+                width: 1.75rem;
             }
         }
